requirements.txt
streamlit
streamlit-option-menu
google-genai
dotenv